Drop support for Spark 3.3
