+++ spark/src/test/scala/org/apache/spark/sql/CometTestBase.scala
@@ -47,7 +47,6 @@ import org.apache.spark.sql.test._
 import org.apache.spark.sql.types.{DecimalType, StructType}
 
 import org.apache.comet._
-import org.apache.comet.CometSparkSessionExtensions.isSpark34Plus
 import org.apache.comet.shims.ShimCometSparkSessionExtensions
 
 /**
@@ -440,114 +439,57 @@ abstract class CometTestBase
       // here: https://github.com/apache/parquet-java/issues/3142
       // here: https://github.com/apache/arrow-rs/issues/7040
       // and here: https://github.com/apache/datafusion-comet/issues/1348
-      if (isSpark34Plus) {
-        """
-         |message root {
-         |  optional boolean                  _1;
-         |  optional int32                    _2(INT_8);
-         |  optional int32                    _3(INT_16);
-         |  optional int32                    _4;
-         |  optional int64                    _5;
-         |  optional float                    _6;
-         |  optional double                   _7;
-         |  optional binary                   _8(UTF8);
-         |  optional int32                    _9(UINT_32);
-         |  optional int32                    _10(UINT_32);
-         |  optional int32                    _11(UINT_32);
-         |  optional int64                    _12(UINT_64);
-         |  optional binary                   _13(ENUM);
-         |  optional FIXED_LEN_BYTE_ARRAY(3)  _14;
-         |  optional int32                    _15(DECIMAL(5, 2));
-         |  optional int64                    _16(DECIMAL(18, 10));
-         |  optional FIXED_LEN_BYTE_ARRAY(16) _17(DECIMAL(38, 37));
-         |  optional INT64                    _18(TIMESTAMP(MILLIS,true));
-         |  optional INT64                    _19(TIMESTAMP(MICROS,true));
-         |  optional INT32                    _20(DATE);
-         |  optional INT32                    _id;
-         |}
-        """.stripMargin
-      } else {
-        """
-         |message root {
-         |  optional boolean                  _1;
-         |  optional int32                    _2(INT_8);
-         |  optional int32                    _3(INT_16);
-         |  optional int32                    _4;
-         |  optional int64                    _5;
-         |  optional float                    _6;
-         |  optional double                   _7;
-         |  optional binary                   _8(UTF8);
-         |  optional int32                    _9(UINT_32);
-         |  optional int32                    _10(UINT_32);
-         |  optional int32                    _11(UINT_32);
-         |  optional int64                    _12(UINT_64);
-         |  optional binary                   _13(ENUM);
-         |  optional binary                   _14(UTF8);
-         |  optional int32                    _15(DECIMAL(5, 2));
-         |  optional int64                    _16(DECIMAL(18, 10));
-         |  optional FIXED_LEN_BYTE_ARRAY(16) _17(DECIMAL(38, 37));
-         |  optional INT64                    _18(TIMESTAMP(MILLIS,true));
-         |  optional INT64                    _19(TIMESTAMP(MICROS,true));
-         |  optional INT32                    _20(DATE);
-         |  optional INT32                    _id;
-         |}
-        """.stripMargin
-      }
+      """
+       |message root {
+       |  optional boolean                  _1;
+       |  optional int32                    _2(INT_8);
+       |  optional int32                    _3(INT_16);
+       |  optional int32                    _4;
+       |  optional int64                    _5;
+       |  optional float                    _6;
+       |  optional double                   _7;
+       |  optional binary                   _8(UTF8);
+       |  optional int32                    _9(UINT_32);
+       |  optional int32                    _10(UINT_32);
+       |  optional int32                    _11(UINT_32);
+       |  optional int64                    _12(UINT_64);
+       |  optional binary                   _13(ENUM);
+       |  optional FIXED_LEN_BYTE_ARRAY(3)  _14;
+       |  optional int32                    _15(DECIMAL(5, 2));
+       |  optional int64                    _16(DECIMAL(18, 10));
+       |  optional FIXED_LEN_BYTE_ARRAY(16) _17(DECIMAL(38, 37));
+       |  optional INT64                    _18(TIMESTAMP(MILLIS,true));
+       |  optional INT64                    _19(TIMESTAMP(MICROS,true));
+       |  optional INT32                    _20(DATE);
+       |  optional INT32                    _id;
+       |}
+      """.stripMargin
     } else {
-
-      if (isSpark34Plus) {
-        """
-         |message root {
-         |  optional boolean                  _1;
-         |  optional int32                    _2(INT_8);
-         |  optional int32                    _3(INT_16);
-         |  optional int32                    _4;
-         |  optional int64                    _5;
-         |  optional float                    _6;
-         |  optional double                   _7;
-         |  optional binary                   _8(UTF8);
-         |  optional int32                    _9(UINT_8);
-         |  optional int32                    _10(UINT_16);
-         |  optional int32                    _11(UINT_32);
-         |  optional int64                    _12(UINT_64);
-         |  optional binary                   _13(ENUM);
-         |  optional FIXED_LEN_BYTE_ARRAY(3)  _14;
-         |  optional int32                    _15(DECIMAL(5, 2));
-         |  optional int64                    _16(DECIMAL(18, 10));
-         |  optional FIXED_LEN_BYTE_ARRAY(16) _17(DECIMAL(38, 37));
-         |  optional INT64                    _18(TIMESTAMP(MILLIS,true));
-         |  optional INT64                    _19(TIMESTAMP(MICROS,true));
-         |  optional INT32                    _20(DATE);
-         |  optional INT32                    _id;
-         |}
-        """.stripMargin
-      } else {
-        """
-         |message root {
-         |  optional boolean                  _1;
-         |  optional int32                    _2(INT_8);
-         |  optional int32                    _3(INT_16);
-         |  optional int32                    _4;
-         |  optional int64                    _5;
-         |  optional float                    _6;
-         |  optional double                   _7;
-         |  optional binary                   _8(UTF8);
-         |  optional int32                    _9(UINT_8);
-         |  optional int32                    _10(UINT_16);
-         |  optional int32                    _11(UINT_32);
-         |  optional int64                    _12(UINT_64);
-         |  optional binary                   _13(ENUM);
-         |  optional binary                   _14(UTF8);
-         |  optional int32                    _15(DECIMAL(5, 2));
-         |  optional int64                    _16(DECIMAL(18, 10));
-         |  optional FIXED_LEN_BYTE_ARRAY(16) _17(DECIMAL(38, 37));
-         |  optional INT64                    _18(TIMESTAMP(MILLIS,true));
-         |  optional INT64                    _19(TIMESTAMP(MICROS,true));
-         |  optional INT32                    _20(DATE);
-         |  optional INT32                    _id;
-         |}
-        """.stripMargin
-      }
+      """
+       |message root {
+       |  optional boolean                  _1;
+       |  optional int32                    _2(INT_8);
+       |  optional int32                    _3(INT_16);
+       |  optional int32                    _4;
+       |  optional int64                    _5;
+       |  optional float                    _6;
+       |  optional double                   _7;
+       |  optional binary                   _8(UTF8);
+       |  optional int32                    _9(UINT_8);
+       |  optional int32                    _10(UINT_16);
+       |  optional int32                    _11(UINT_32);
+       |  optional int64                    _12(UINT_64);
+       |  optional binary                   _13(ENUM);
+       |  optional FIXED_LEN_BYTE_ARRAY(3)  _14;
+       |  optional int32                    _15(DECIMAL(5, 2));
+       |  optional int64                    _16(DECIMAL(18, 10));
+       |  optional FIXED_LEN_BYTE_ARRAY(16) _17(DECIMAL(38, 37));
+       |  optional INT64                    _18(TIMESTAMP(MILLIS,true));
+       |  optional INT64                    _19(TIMESTAMP(MICROS,true));
+       |  optional INT32                    _20(DATE);
+       |  optional INT32                    _id;
+       |}
+      """.stripMargin
     }
   }
