+++ spark/src/test/spark-3.4-plus/org/apache/comet/exec/CometExec3_4PlusSuite.scala
@@ -19,12 +19,18 @@
 
 package org.apache.comet.exec
 
+import java.io.ByteArrayOutputStream
+
+import scala.util.Random
+
+import org.apache.spark.sql.{Column, CometTestBase}
+import org.apache.comet.CometConf
+import org.apache.spark.sql.catalyst.expressions.BloomFilterMightContain
+import org.apache.spark.sql.functions.{col, lit}
+import org.apache.spark.util.sketch.BloomFilter
+
 import org.scalactic.source.Position
 import org.scalatest.Tag
-
-import org.apache.spark.sql.CometTestBase
-import org.apache.comet.CometConf
-
 /**
  * This test suite contains tests for only Spark 3.4+.
  */
@@ -100,4 +106,62 @@ class CometExec3_4PlusSuite extends CometTestBase {
       checkSparkAnswer(mapData.toDF().offset(99))
     }
   }
+
+  test("test BloomFilterMightContain can take a constant value input") {
+    val table = "test"
+
+    withTable(table) {
+      sql(s"create table $table(col1 long, col2 int) using parquet")
+      sql(s"insert into $table values (201, 1)")
+      checkSparkAnswerAndOperator(
+        s"""
+           |SELECT might_contain(
+           |X'00000001000000050000000343A2EC6EA8C117E2D3CDB767296B144FC5BFBCED9737F267', col1) FROM $table
+           |""".stripMargin)
+    }
+  }
+
+  test("test NULL inputs for BloomFilterMightContain") {
+    val table = "test"
+
+    withTable(table) {
+      sql(s"create table $table(col1 long, col2 int) using parquet")
+      sql(s"insert into $table values (201, 1), (null, 2)")
+      checkSparkAnswerAndOperator(
+        s"""
+           |SELECT might_contain(null, null) both_null,
+           |       might_contain(null, 1L) null_bf,
+           |       might_contain(
+           |         X'00000001000000050000000343A2EC6EA8C117E2D3CDB767296B144FC5BFBCED9737F267', col1) null_value
+           |       FROM $table
+           |""".stripMargin)
+    }
+  }
+
+  test("test BloomFilterMightContain from random input") {
+    val (longs, bfBytes) = bloomFilterFromRandomInput(10000, 10000)
+    val table = "test"
+
+    withTable(table) {
+      sql(s"create table $table(col1 long, col2 binary) using parquet")
+      spark.createDataset(longs).map(x => (x, bfBytes)).toDF("col1", "col2").write.insertInto(table)
+      val df = spark.table(table).select(new Column(BloomFilterMightContain(lit(bfBytes).expr, col("col1").expr)))
+      checkSparkAnswerAndOperator(df)
+      // check with scalar subquery
+      checkSparkAnswerAndOperator(
+        s"""
+           |SELECT might_contain((select first(col2) as col2 from $table), col1) FROM $table
+           |""".stripMargin)
+    }
+  }
+
+  private def bloomFilterFromRandomInput(expectedItems: Long, expectedBits: Long): (Seq[Long], Array[Byte]) = {
+    val bf = BloomFilter.create(expectedItems, expectedBits)
+    val longs = (0 until expectedItems.toInt).map(_ => Random.nextLong())
+    longs.foreach(bf.put)
+    val os = new ByteArrayOutputStream()
+    bf.writeTo(os)
+    (longs, os.toByteArray)
+  }
+
 }
